% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dnds_load_annotation}
\alias{dnds_load_annotation}
\title{A DndsAnalysis Function}
\usage{
dnds_load_annotation(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
dnds_load_annotation()
}
\keyword{cats}

